Revert "ssss"
This reverts commit 18dd352007a0da4afb027185bc22e5aef72cadca.

+++ RentACarApiUser/src/main/java/com/appsdevelopersblog/rentacar/api/user/business/concretes/CorporateCustomerManager.java
@@ -77,10 +77,10 @@ public class CorporateCustomerManager implements CorporateCustomerService {
 
 	@Override
 	public Result delete(DeleteCorporateCustomerRequest deleteCorporateCustomerRequest) {
-//		Result result = BusinessRules.run(checkCorporateCustomerExistsBuId(deleteCorporateCustomerRequest.getCustomerId()));
-//		if (result != null) {
-//			return result;
-//		}
+		Result result = BusinessRules.run(checkCorporateCustomerExistsBuId(deleteCorporateCustomerRequest.getCustomerId()));
+		if (result != null) {
+			return result;
+		}
 		CorporateCustomer corporateCustomer = modelMapperService.forRequest().map(deleteCorporateCustomerRequest,
 				CorporateCustomer.class);
 		this.corporateCustomerDao.delete(corporateCustomer);
